Refactoring scripts
  Separated setup.sh into config.sh and functions.sh. In config.sh,
it support some variables, but it would be not enough yet. In
functions.sh, the platform that this script supports is ubuntu.

The built image of AOSP is send to "OUTPUT_DIST" variable in config.sh.
Currently it supports S3 via s3cmd

We can execute run.sh on ubuntu platform.